addressing spham's feedback
