+++ scripts/o3de/tests/ExportScripts/CMakeLists.txt
@@ -25,7 +25,7 @@ ly_add_pytest(
 )
 
 ly_add_pytest(
-    NAME o3de_export_source_ios_xcode
+    NAME o3de_export_source_android
     PATH ${CMAKE_CURRENT_LIST_DIR}/test_export_source_android.py
     TEST_SUITE smoke
     EXCLUDE_TEST_RUN_TARGET_FROM_IDE
